+++ packages/backend/test-federation/test/note.test.ts
@@ -1,7 +1,6 @@
 import assert, { rejects, strictEqual } from 'node:assert';
-import strict from 'node:assert/strict';
 import * as Misskey from 'misskey-js';
-import { addCustomEmoji, createAccount, createModerator, deepStrictEqualWithExcludedFields, fetchAdmin, type LoginUser, resolveRemoteNote, resolveRemoteUser, sleep, uploadFile } from './utils.js';
+import { addCustomEmoji, createAccount, createModerator, deepStrictEqualWithExcludedFields, type LoginUser, resolveRemoteNote, resolveRemoteUser, sleep, uploadFile } from './utils.js';
 
 describe('Note', () => {
 	let alice: LoginUser, bob: LoginUser;
@@ -377,58 +376,6 @@ describe('Note', () => {
 				strictEqual(noteAfterVote.poll.choices[0].votes, 1);
 				strictEqual(noteAfterVote.poll.choices[1].votes, 0);
 			});
-		});
-	});
-
-	describe('Reacted Remote Note', () => {
-		let cAdmin: LoginUser;
-		let carol: LoginUser;
-		let bobInC: Misskey.entities.UserDetailedNotMe;
-
-		beforeAll(async() => {
-			cAdmin = await fetchAdmin('c.test');
-
-			await cAdmin.client.request('admin/update-meta', {
-				resolveRemoteReactedNotes: true,
-			});
-
-			await sleep(1000);
-
-			carol = await createAccount('c.test');
-
-			await sleep(1000);
-
-			bobInC = await resolveRemoteUser('b.test', bob.id, carol);
-		});
-
-		test('Exist of alice note reacted by bob in c.test', async () => {
-			await carol.client.request('following/create', { userId: bobInC.id });
-			await sleep(1000);
-
-			const note = (await alice.client.request('notes/create', { text: 'test note from alice' })).createdNote;
-			const noteInB = await resolveRemoteNote('a.test', note.id, bob);
-			await bob.client.request('notes/reactions/create', { noteId: noteInB.id, reaction: '❤' });
-
-			// Wait for ActivityPub delivery and note resolution with retry logic
-			const expectedUri = `https://a.test/notes/${note.id}`;
-			let reactedNoteInC: Misskey.entities.Note | undefined;
-			for (let i = 0; i < 10; i++) {
-				await sleep(1000);
-				const notesInC = await carol.client.request('notes/global-timeline', {});
-				reactedNoteInC = notesInC.find(n => n.uri === expectedUri);
-				if (reactedNoteInC) break;
-			}
-
-			assert(reactedNoteInC != null, 'Note should be resolved in c.test after Bob reacted to it');
-			strictEqual(reactedNoteInC.text, note.text);
-			strictEqual(reactedNoteInC.createdAt, note.createdAt);
-		});
-
-		afterAll(async() => {
-			await cAdmin.client.request('admin/update-meta', {
-				resolveRemoteReactedNotes: false,
-			});
-			await sleep();
 		});
 	});
 });
